Fix 2 issues (EnumKey length = 1) + Loading same widget multiple times
Fix for Enumeration keys with a length of 1 (these were ignored because of length > 1 instead of length >=1)
Fix for label "for" and input id's where the widget was loaded more than once on a page. These id's weren't unique anymore. Fixed by adding the widget-id to these for / id attributes.